Print when we are sorting the results
When there are many hundreds of thousands of files to consider,
and especially when all of the dentries and inodes are in the
page cache already, sorting the results can take just as long as
printing them.

Don't give the false impression that the last action that was
performed (e.g. mapping the directory argument given) is still
being performed when we're actually sorting the results prior to
printing them.